changes added for Put api

+++ index.js
@@ -1,12 +1,13 @@
 const form = document.getElementById('formId');
 const detailsDiv = document.getElementById('details');
 
-// on refresh get data from crudCrud and Using showOutput() to add li tag on page
+// on Page-Reloading(refresh) it will get data from crudCrud and Using showOutput() to add li tag on page
 window.addEventListener('DOMContentLoaded',()=>{
   axios
-    .get('https://crudcrud.com/api/47feaa5cf936419abd3dc565403108c5/appointment')
+    .get('https://crudcrud.com/api/e496ba72a8c24acaab76cb4aa277f33a/appointment')
     .then((res)=>{
       for (let index = 0; index < res.data.length; index++) {
+        // to add each data on html page as listItem
          showOutput(res.data[index]);
       }
       console.log(res.data);
@@ -14,6 +15,7 @@ window.addEventListener('DOMContentLoaded',()=>{
     .catch( (err)=> console.log(err))
 })
 
+// this fuction will add listItems on html page
 function showOutput(object){
   const list = document.createElement('li');
   list.textContent = `${object.name} - ${object.email} - ${object.phone}`;
@@ -36,6 +38,7 @@ function showOutput(object){
 
 form.addEventListener('submit', onSubmit);
 
+// submit fun will take the formData & performs PUT or POST based on conditions
 function onSubmit(e) {
   e.preventDefault();
 
@@ -54,14 +57,45 @@ function onSubmit(e) {
     phone: phoneNumber,
   };
 
+  // below axios will checks for the newly inserted form-details in server
+  // if it is present in server then calls PUT api else calls POST api
+  // i have used flagPost boolean variable for api-call toggling b/w the POST & PUT 
+  var flagPost=true;
+  var putId;
+  axios
+    .get('https://crudcrud.com/api/e496ba72a8c24acaab76cb4aa277f33a/appointment')
+    .then((res)=>{
+      for (let index = 0; index < res.data.length; index++) {
+        // the if() checks the data is already present in server or not  if it's there then calls Put api's function
+        if(res.data[index].name==formData.name || res.data[index].email == formData.email){
+            flagPost=false;
+            putId=res.data[index]._id;
+            updateDetails(formData);
+        }
+      }
+      // this if-block is to call Post-api using the flagPost toggler if the form-data is not on server
+      if(flagPost==true){
+        saveToCrud();
+      }
+    })
+    .catch((err)=>console.log(err));
+
+  function updateDetails(obj){
+    // update's if data is present in the server
+    axios.put('https://crudcrud.com/api/e496ba72a8c24acaab76cb4aa277f33a/appointment/'+putId,obj)
+    .then((resp)=>console.log("updated succesly",resp))
+    .catch((er)=>console.log(er))
+  }
+
   function saveToCrud(){
+    //post's if the data is not in the server
     axios
-      .post('https://crudcrud.com/api/47feaa5cf936419abd3dc565403108c5/appointment',formData)
+      .post('https://crudcrud.com/api/e496ba72a8c24acaab76cb4aa277f33a/appointment',formData)
       .then((res)=>console.log(res.data))
       .catch( (err)=> console.log(err));
-
+      
   }
-  saveToCrud();
+  
   // showOutput will create li tag and add data on html page
   showOutput(formData);
   // Clear input fields after submission
@@ -72,36 +106,44 @@ function onSubmit(e) {
 
 var del=document.getElementById('details');
 del.addEventListener('click',deleteItem);
-// to delete based on id 
-// first using GET to get id 
-// second using DELETE to delete based on matched id
+
+// This function will delete data from server if clicked on delete btn only
+// if edit btn is clicked it will delete data from html page only not from server
 function deleteItem(e){
     let parent=e.target.parentElement;
     var array=e.target.parentElement.textContent.split('-');
-    // console.log("array is",array);
     var urlId;
-    // GET to get id by using email via traversing the obj's array
-    axios
-    .get('https://crudcrud.com/api/47feaa5cf936419abd3dc565403108c5/appointment')
-    .then((res)=>{
-      // console.log("inside get axios");
-      for (let index = 0; index < res.data.length; index++) {
-        if(res.data[index].email==array[1].trim()){
-            // console.log("inside get axios1");
-            urlId=res.data[index]._id;
-            // console.log("url obj id is:",urlId);
-         }
-      }
-      // // to delete from crudcrud
-      const options={
-        method:'delete',
-        url:'https://crudcrud.com/api/47feaa5cf936419abd3dc565403108c5/appointment/'+urlId,
-      }
-      axios(options)
-      .then((res)=>console.log(res,"deleted successfully"))
-      .catch((err)=>console.log(err));
-    })
-    .catch((err)=>console.log(err));
+    // to delete listItem on html page
     del.removeChild(parent);
-    
+
+    if(e.target.id==='edit'){
+      document.getElementById('username').value=array[0].trim();
+      document.getElementById('email').value=array[1].trim();
+      document.getElementById('number').value=array[2].trim();
+    }   
+    else{
+      // to delete data from server based on id 
+      // first using GET to get id 
+      // second using DELETE to delete based on matched id
+      axios
+      .get('https://crudcrud.com/api/e496ba72a8c24acaab76cb4aa277f33a/appointment')
+      .then((res)=>{
+        // GET to get id by using email via traversing the obj's array
+        for (let index = 0; index < res.data.length; index++) {
+          if(res.data[index].email==array[1].trim()){
+            // collecting id that will be passed to delete api's url
+              urlId=res.data[index]._id;
+          }
+        }
+        // // to delete from crudcrud
+        const options={
+          method:'delete',
+          url:'https://crudcrud.com/api/e496ba72a8c24acaab76cb4aa277f33a/appointment/'+urlId,
+        }
+        axios(options)
+        .then((res)=>console.log(res,"deleted successfully"))
+        .catch((err)=>console.log(err));
+      })
+      .catch((err)=>console.log(err));
+    }
 }
